Move extra coords into it's own dir for splitting
