+++ ndcube/extra_coords/__init__.py
@@ -1,0 +1,4 @@
+from .extra_coords import ExtraCoords
+from .lookup_table_coord import LookupTableCoord
+
+__all__ = ['ExtraCoords', 'LookupTableCoord']
